physics: Various performance improvements.
Allocation Elimination (Hot Path)
1. Math.Sign() — zero-alloc bit test (Common/Math.cs)
   Replaced a NativeArray allocation per call with math.asint() bit manipulation. This was called ~2.5M times/sec from narrow-phase collision detection.
2. PhysicsDynamicBroadPhase.FindOverlaps — stack-allocated removal buffer (Game/PhysicsDynamicBroadPhase.cs)
   Replaced ToNativeArray(TempJob) with FixedList64Bytes<int> for collecting hash set removals during octree overlap queries. Eliminates a temp allocation per ball pair check.
3. InsideOfs.GetBitIndex() — bitmask free-bit find (Game/InsideOfs.cs)
   Replaced GetValueArray(Allocator.TempJob) + O(64*n) Contains() scan with ulong bitmask + math.tzcnt() for O(1) free-bit lookup. Zero allocations.
4. InsideOfs.GetIdsOfBallsInsideItem() — stack return type (Game/InsideOfs.cs)
   Changed return from List<int> (heap) to FixedList64Bytes<int> (stack, up to 15 ints). Updated all callers (PhysicsEngine, BumperApi, KickerColliderComponent).
Persistent Native Container Reuse
5. Persistent ball octree (PhysicsEngineContext.BallOctree)
   Moved NativeOctree<int> for ball-to-ball collisions from per-tick TempJob alloc/dispose to a persistent field. Clear() + re-insert each cycle instead of create/destroy.
6. Persistent kinematic octree with dirty flag (PhysicsEngineContext.KinematicOctree)
   Same pattern, plus only rebuilt when KinematicOctreeDirty is set (transforms actually changed). Reduces rebuild frequency from ~1kHz to ~60Hz.
7. Persistent PhysicsCycle (PhysicsEngineContext.PhysicsCycle)
   Moved from per-tick creation to persistent field. Avoids per-tick allocation of the internal NativeList<ContactBufferElement>.
Redundant Work Reduction
8. InsideOfs double-lookup elimination (Game/InsideOfs.cs)
   Replaced ContainsKey + [] patterns with TryGetValue in SetOutsideOfAll, IsInsideOf, GetInsideCount, IsEmpty, GetIdsOfBallsInsideItem.
9. PhysFactor float conversion (Engine/Common/Constants.cs + 6 files)
   Changed PhysFactor from double to float, removing 8 redundant (float) casts. Avoids implicit double-precision promotion in Burst-compiled velocity code.
10. VPOrthonormalize float3 conversion (VPT/Ball/BallDisplacementPhysics.cs)
    Replaced UnityEngine.Vector3 with Unity.Mathematics.float3 + math.cross()/math.normalize(). Eliminates Vector3↔float3 conversions on every ball displacement.
Safety
11. Max substep cap (PhysicsUpdate.cs)
    Added PhysicsConstants.MaxSubSteps = 200. If physics falls behind (e.g. frame hitch), the loop caps iterations and skips time forward instead of cascading into a death spiral.
---
Net effect: The physics tick (running at 1kHz) is now allocation-free on its hot path. The three largest per-tick native allocations (two octrees + contacts list) are gone, replaced with persistent containers that clear and refill. The remaining improvements reduce redundant hash lookups and eliminate unnecessary type conversions in Burst-compiled code.

+++ VisualPinball.Unity/VisualPinball.Unity/VPT/Ball/BallVelocityPhysics.cs
@@ -37,7 +37,7 @@ namespace VisualPinball.Unity
 					-2.0f
 				);
 			} else {
-				ball.Velocity += (float)PhysicsConstants.PhysFactor * gravity;
+				ball.Velocity += PhysicsConstants.PhysFactor * gravity;
 			}
 		}
 	}
